requirements.txt
Flask==3.0.3
gunicorn==22.0.0
playwright==1.47.2
python-dateutil==2.9.0.post0
